Fully added resource generation and consumption to buildings. Changed shallow copy into deep copy for resourceTypes.
Fully added resource generation and consumption to buildings. But need to add info to all buildings and to fix the building names as they have no spaces.
Changed shallow copy into deep copy for storage because it was using a spread of resourceTypes but now it uses json parse for a deep copy.