+Unclaimed Cell Menu Functionality +Minor Validiation
+Unclaimed Cell Menu Functionality +Minor Validiation.
Made it so the claim button appears on all unclaimed cells instead of only those eligible to be claimed but the button is disabled.
Set the foundation for making a menu for the first cell to claim in the beginning of the game to be separate from the Unclaimed Cell Menu.
Added minor validation by throwing an error on interaction with any cell somehow neither claimed nor unclaimed.