+Building construction time +Buildings cell feature more detailed with unique ID
New object added to represent each individual building on a cell as part of cell features instead of a single name. This new object currently contains the building name, construction timer and unique ID. The construction timer is tracked here but its data is from buildingData. Unique ID is based on a simple iteration so each building added will be +1 to what the last number was.